requirements.txt
aiohttp==3.12.13
openai==1.93.0
python-telegram-bot==22.1
